requirements.txt
scapy
mysql-connector-python
python-dotenv
matplotlib
pandas
ttkbootstrap